+++ packages/pwa-kit-dev/src/configs/jest/jest.config.js
@@ -7,14 +7,12 @@
 import path from 'path'
 
 module.exports = {
-    testURL: 'http://localhost/',
     verbose: true,
     collectCoverage: true,
-    // We need access to jsdom globally in tests.
-    // jsdom isn't accessible so we need to use this
-    // 3rd party test environment wrapper. When we
-    // upgrade to jest 28, we can revert back to jsdom.
     testEnvironment: 'jest-environment-jsdom-global',
+    testEnvironmentOptions: {
+        url: 'http://localhost/'
+    },
     testPathIgnorePatterns: ['node_modules', 'build'],
     moduleNameMapper: {
         '\\.(jpg|jpeg|png|gif|eot|otf|webp|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':