+++ packages/template-retail-react-app/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v9.0.0-preview.1 (Feb 09, 2026)
 - [Bugfix] 1CC Bug Fixes [#3638](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3638)
+- [Bugfix] 1CC UX changes: update edit labels to "Change", fix billing address display logic [W-21240479]
 
 ## v9.0.0-preview.0 (Feb 06, 2026)
 - [Feature] One Click Checkout [#3552](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3552)