+++ packages/template-retail-react-app/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## v9.0.0-dev (Jan 19, 2026)
-- Update jest-fetch-mock from ^2.1.2 to ^3.0.3
+- Update jest-fetch-mock [#3663](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3663)
 - Add Node 24 support. Drop Node 16 support [#3652](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3652)
 - [Feature] One Click Checkout [#3552](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3552)
 - [Feature] Add `fuzzyPathMatching` to reduce computational overhead of route generation at time of application load [#3530](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3530)